requirements.txt
setuptools==67.0.0
sanic==22.12.0
sanic-ext==22.12.0
tortoise-orm==0.19.3
asyncmy==0.2.5
aiomysql==0.1.1
asyncpg==0.27.0
sanic-testing==22.12.0
pytest==7.2.1
# pytest-xdist==3.1.0